requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.2
azure-cognitiveservices-speech==1.34.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.1.0
boto3==1.33.8
botocore==1.33.13
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
celery-redbeat==2.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
distro==1.9.0
Django==5.0
django-celery-results==2.5.1
django-compat==1.0.15
django-constance==2.9.1
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-filter==23.5
django-multiselectfield==0.1.13
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
drf-yasg==1.21.8
ecdsa==0.19.0
enum-compat==0.0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.16.1
firebase-admin==6.5.0
flake8==7.0.0
flower==2.0.0
fpdf==1.7.2
fsspec==2024.9.0
future==1.0.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.4
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.66.2
grpcio-health-checking==1.62.3
grpcio-status==1.62.3
grpcio-tools==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.1
humanize==4.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
intuit-oauth==1.2.4
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
lxml==5.3.0
MarkupSafe==2.1.5
mccabe==0.7.0
mock==5.1.0
mpmath==1.3.0
msgpack==1.1.0
mypy==1.8.0
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==2.0.0
oauthlib==3.2.2
openai==1.35.13
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.1
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyflakes==3.2.0
PyJWT==2.8.0
pymsteams==0.2.2
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.2
pypdf==4.3.1
PySocks==1.7.1
pyTelegramBotAPI==4.21.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==0.13.0
python-http-client==3.3.7
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.5
selenium==4.21.0
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==10.10.0
sympy==1.13.3
tenacity==8.2.3
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
tornado==6.4
tqdm==4.66.4
transformers==4.43.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.6
validators==0.33.0
vine==5.1.0
wcwidth==0.2.13
weaviate-client==4.7.1
websockets==12.0
wsproto==1.2.0
google-cloud-vision==3.9.0
tiktoken==0.8.0
langdetect==1.0.9